Address review feedback: dedupe PipeBuffer, harden follow-test, elaborate signal comments
Review findings from PR #133 that are worth landing rather than
deferring:

- Extract `PipeBuffer` (née StderrBuffer) to its own file so both
  RunCommandTests and the new LogsCommandTests share one
  thread-safe pipe accumulator.
- Replace the 300ms magic sleep in the follow-mode test with a
  poke-on-every-poll pattern: rewrite the sentinel on each iteration
  until the stream observes it. Removes the dependency on
  subprocess-startup timing.
- Expand the LogsCommand signal-handling comments to document (a)
  the narrow benign race between process.run() and signal(SIG_IGN)
  and (b) why restoring SIG_DFL rather than the prior disposition
  is safe for this leaf command.
- Call out in the README that `previewsmcp logs` honors
  `PREVIEWSMCP_SOCKET_DIR` only when the variable is actually set
  for the `logs` invocation — an easy footgun when switching
  terminals during a debug session.
